+++ src/conditions.c
@@ -16,6 +16,33 @@ void vec_assert(SEXP x, struct vctrs_arg* arg) {
   }
 }
 
+// [[ include("vctrs.h") ]]
+void stop_incompatible_type(SEXP x,
+                            SEXP y,
+                            struct vctrs_arg* x_arg,
+                            struct vctrs_arg* y_arg) {
+  SEXP syms[5] = {
+    syms_x,
+    syms_y,
+    syms_x_arg,
+    syms_y_arg,
+    NULL
+  };
+  SEXP args[5] = {
+    PROTECT(r_protect(x)),
+    PROTECT(r_protect(y)),
+    PROTECT(vctrs_arg(x_arg)),
+    PROTECT(vctrs_arg(y_arg)),
+    NULL
+  };
+
+  SEXP call = PROTECT(r_call(syms_stop_incompatible_type, syms, args));
+  Rf_eval(call, vctrs_ns_env);
+
+  never_reached("stop_incompatible_type");
+}
+
+// [[ include("vctrs.h") ]]
 void stop_incompatible_size(SEXP x, SEXP y,
                             R_len_t x_size, R_len_t y_size,
                             struct vctrs_arg* x_arg,
@@ -23,8 +50,8 @@ void stop_incompatible_size(SEXP x, SEXP y,
   SEXP syms[7] = {
     syms_x,
     syms_y,
-    r_sym("x_size"),
-    r_sym("y_size"),
+    syms_x_size,
+    syms_y_size,
     syms_x_arg,
     syms_y_arg,
     NULL
@@ -39,7 +66,7 @@ void stop_incompatible_size(SEXP x, SEXP y,
     NULL
   };
 
-  SEXP call = PROTECT(r_call(r_sym("stop_incompatible_size"), syms, args));
+  SEXP call = PROTECT(r_call(syms_stop_incompatible_size, syms, args));
   Rf_eval(call, vctrs_ns_env);
 
   never_reached("stop_incompatible_size");