+++ NEWS.md
@@ -215,6 +215,8 @@ The following errors are caused by breaking changes.
 
 * `POSIXlt` and `POSIXct` vectors are handled more consistently (#901).
 
+* Ordered factors that do not have identical levels are now incompatible.
+
 
 ## Indexing and names